requirements.txt
fastapi>=0.104.0,<0.110.0
uvicorn[standard]>=0.24.0,<0.28.0
pymongo>=4.6.0,<5.0.0
motor>=3.3.0,<4.0.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
email-validator>=2.1.1
bcrypt>=4.1.0
Pillow>=10.0.0
aiofiles>=23.2.0
